fix/154: change Song::time to be a 32 bit integer
Increase the width of the track length field to allow longer song lengths.
32 bits is enough.

+++ mpd-interface/cuefile.cpp
@@ -575,7 +575,7 @@ bool CueFile::parse(const QString& fileName, const QString& dir, QList<Song>& so
 			if (!indexTime(tracks.at(i).value("index"), tracks[i + 1].value("index"), time)) {
 				continue;// if incorrect, then skip the track (jump to the next one)
 			}
-			song.time = static_cast<quint16>(time);
+			song.time = static_cast<quint32>(time);
 		}
 		else {
 			// incorrect index?
